requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9  # If using PostgreSQL, else omit for SQLite
gunicorn>=20.1.0      # For deployment (optional)
